requirements.txt
igdb_api_v4==0.0.5
python-dotenv==0.20.0
